+++ packages/autometrics-typescript-plugin/src/astHelpers.ts
@@ -1,7 +1,6 @@
 import ts from "typescript/lib/tsserverlibrary";
 
 import type { NodeType } from "./types";
-import { hasAutometricsDecorator } from "./utils";
 
 /**
  * Checks if the node is wrapped or decorated by Autometrics (but not the
@@ -49,6 +48,19 @@ export function isAutometricsWrappedOrDecorated(
   const isDecorated =
     hasAutometricsDecorator(method) || hasAutometricsDecorator(method.parent);
   return isDecorated;
+}
+
+export function hasAutometricsDecorator(node: ts.Node) {
+  const decorators = ts.canHaveDecorators(node) && ts.getDecorators(node);
+  if (!decorators) {
+    return false;
+  }
+
+  const hasAutometricsDecorator = decorators.some((decorator) =>
+    decorator.getText().startsWith("@Autometrics"),
+  );
+
+  return hasAutometricsDecorator;
 }
 
 /**
